Handle mixed scenarios
We have multiple FTDI devices some of which are relay boards
while others are serial interfaces.

Only devices that we actually use as relay boards shall be detached
from the kernel driver and switched to bitbang mode.

The status of a board is only output if requested or switched.

In examples/relctl.py two new command line options are introduced:

    -g display status
    -k attach to kernel driver

Signed-off-by: Heinrich Schuchardt <xypron.glpk@gmx.de>